Устранено замечение по Д5. Фильтрация объявлений вынесена в новый файл
