+++ js/main.js
@@ -2,3 +2,4 @@ import './map.js';
 import './slider.js';
 import './form-reset.js';
 import './photo-preview.js';
+import './map-filter.js';
